renamed configSlice to global, moved common params to it, update links, add parametric message to LoadingPage
