+++ src/useGetConfig.js
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
-import { setConfig } from "./configSlice";
+import { setConfig } from "./globalSlice";
 
 export const useGetConfig = () => {
     const apiURL = "https://api.themoviedb.org/3/configuration?api_key=768f7875782193f5e4797762314da0b7";
